refactor: redundancy fixes
