+++ src/commands/fun/selfmute.ts
@@ -1,7 +1,7 @@
 import type { GuildCommandContext } from "seyfert";
 import { Command, createStringOption, Declare, Embed, Options } from "seyfert";
 import { EmbedColors } from "seyfert/lib/common";
-import ms from "@/utils/ms";
+import { isValid, parse } from "@/utils/ms";
 
 const options = {
   time: createStringOption({
@@ -19,7 +19,7 @@ export default class SelfMuteCommand extends Command {
   async run(ctx: GuildCommandContext<typeof options>) {
     const { time } = ctx.options;
 
-    if (!ms.isValid(time))
+    if (!isValid(time))
       return await ctx.write({
         content:
           "✗ Formato de tiempo invalido. **Ejemplos válidos:** 10min, 1h, 3d, 2m, 5s.",
@@ -31,7 +31,7 @@ export default class SelfMuteCommand extends Command {
         content: "✗ No tengo los permisos suficientes.",
       });
 
-    const milliseconds = ms.parse(time) || 0;
+    const milliseconds = parse(time) || 0;
     ctx.member.timeout(milliseconds, `Comando self-mute | Tiempo: ${time}`);
 
     const successEmbed = new Embed({
