+++ src/commands/moderation/warn/parent.ts
@@ -1,12 +1,8 @@
-import { Command, Declare, Options } from "seyfert";
-
-import { AddCommand } from "./add.command";
-import { ListCommand } from "./list.command";
-import { RemoveCommand } from "./remove.command";
+import { AutoLoad, Command, Declare } from "seyfert";
 
 @Declare({
   name: "warn",
   description: "Manejar los warns de los usuarios",
 })
-@Options([AddCommand, RemoveCommand, ListCommand])
-export default class AccountCommand extends Command {}
+@AutoLoad()
+export default class WarnParent extends Command {}
