+++ src/db.ts
@@ -1,8 +1,5 @@
-import dotenv from "dotenv";
 import { drizzle } from "drizzle-orm/node-postgres";
 import pg from "pg";
-
-dotenv.config();
 
 const pool = new pg.Pool({
   connectionString: process.env.DATABASE_URL,
