+++ src/repositories/index.ts
@@ -1,5 +1,5 @@
-import { userRepository } from "./userRepository";
+import * as user from "./user";
 
 export const repositories = {
-  user: userRepository,
+  user,
 };
