+++ src/schemas/user.ts
@@ -1,18 +1,18 @@
 import { integer, jsonb, pgTable, varchar } from "drizzle-orm/pg-core";
 
-export type User = {
+export interface User {
   id: string;
   bank: number;
   cash: number;
   warns: Warn[] | null;
-};
+}
 
-export type Warn = {
+export interface Warn {
   reason: string;
   warn_id: number; // incremental
   moderator: string;
   timestamp: string; // Date ISO
-};
+}
 
 export const users = pgTable("users", {
   id: varchar("id", { length: 50 }).primaryKey().notNull().unique(),
